requirements.txt
# Browser automation
browser-use>=0.1.0

# LLM and API
openai>=1.0.0
pydantic>=2.0.0

# Environment management
python-dotenv>=1.0.0

# HTTP client (for NVIDIA NIM API)
httpx>=0.24.0
